One shot map input to output with logging
This change adds a one-shot command that reads from the tank
input. The input is then mapped to an actual value before being
mapped again to the gauge output. All values are logged in the
console output.

Tests excercise both maps by checking the logged output for the
actual value (which should be linear) as well as the values read
fro the tank and output to the gauge.

Tests:
 - Run all unit tests