Implement map modification
This adds support for editing individual values stored in bins of the
input ("i") or output ("o") maps. The parsing of values is very
tolerant as anything that can be contained in a 16-bit value is valid.
The bin number can be between 0 and 8.

Tests updated to verify new functionality and a trivial error in the
size used by the MapLoadAndSave has been fixed.

Tests
 - Run unit tests under valgrind