Remove printf dependency
Remove the printf dependency by adding a simpler print text and
newline HAL. Two simple routines are added to format uint16_t
values and bin numbers. The buffers for these are added on the
stack which may require changing to be static depending on how
XC8 reacts.

The newlines in the test case have been removed as these are
now included in the (simpler) data structure.

Tests:
 - Run updated unit tests