Refactor to remove scanf from command processing
This change removes the use of scanf from command processing
and replaces with a simple value and bin parser. This removes any dependency on stdio.h which should reduce the code and
data usage when built for the PIC.

An additional success test case has been added for the g
command.

Tests:
 - Run all unit tests