Basic schematic design for a PIC12F1840 based fuel gauge
This is an outline design for a PIC based fuel gauge. It uses a simple current source (LM317 or discrete) to measure the
resistance of a fuel sender (0-100R). The PWM output of the MCU
is output through a MOSFET driver and power MOSFET from
https://tahmidmc.blogspot.com/2012/12/low-side-mosfet-drive-circuits-and_23.html

The changeset only contains a schematic and is intended for
breadboarding at this point.