Feature complete HW design with hack PCB layout
This completes the HW after testing on a breadboard. Both the
sender input and gauge output have been tested.

A fuel low output has been hacked together though this has no
software currently and may be replaced with a 12V indicator.

The PCB layout is functional and passes DRC but is pretty awful
otherwise.

Tests:
 - Run clean run of ERC on the schematics and DRC on the PCB.