Reroute to a smaller 49x81mm board
This re-lays out the board to a much smaller form factor with more intelligent routing. The track size for power nets is reduced
to 1.2mm which should be good for 2.7A max current. A number
of ground plane vias have been thrown in with little thought to
achieve better coverage.