Version 1.0 hardware
This change contains what is hoped to be 1.0 hardware. It removes the complex MOSFET driver for a simple direct drive
from the MCU. The LED has been removed as this appears to
provide little value as the VDO test guage uses 65% to 98% duty
cycle to cover 0% to 100% which is virtually imperceptible.

A simple transistor driver for a low fuel warning LED or lamp has
been included. This should work with a 12V LED indicator lamp.

Tests:
 - Run a full ERC on schematic and DRC on the PCB
 - Check mechanical fit with a 1:1 printout against the Hammond
   1691ABK case
 - Check new MOSFET driver with VDO fuel gauge and calibrate
   output map. MOSFET turn on time seems to be about 1uS
   which should be OK given low current.