Provide a basic set of project documentation
This includes a bit of basic project oducmentation that covers the
motivation, how the device works and little bit of information on the
hardware and software. Much still left to write.

A FreeCAD model of the Ultima fuel tank is included along with a
render. A render of the PCB from Kicad is also included.

Tests:
 - Proof read in Markdown view in Firefox