Update pthread link order
This fixes an ordering issue when linking with pthread by using the
mechanism defined in CMake 3.1 to configure the link order for the pthreads library found on the system. This allows the code to build on Fedora 30 (where link order isn't as fussy) and Ubuntu Xenial (which is fussy).

Tests:
 - Clean build and test run on Fedora 30
 - Clean build and test run on Ubuntu Xenial Docker container