Implement a Low Fuel Warning Light
This change implements a low fuel warning light. The light is
attached to pin6 of the MCU and is activated by pulling this GPIO
pin low.

A low fuel level setting is maintained in EEPROM alongside the
input and output maps. It is configured using the new "f"
command which is similar to the existing "g" gauge output
command.

The fuel level is checked each time the gauge main loop is run.
The actual linear fuel level is the valu that is checked.

A new set of unit tests are included which fully excercise the
loading and storing of the fule level setting.

Fix a number of misspellings of gauge in code comments.

Tests:
 - Run the unit tests
 - Test on real HW with 0 and ffff low fuel limits and verify the
   light is never displayed and always displayed.
 - Test that a set value of 1000 applies at the correct input with a
   pot and linear input map.