Add a mode for continously logging gauge values as it runs
This adds a mode which enables the gauge to continously log
tank input, actual fuel level and gauge output values as the
gauge runs. This is the same output as achieved using the "t"
one-shot command. The new mode is enabled by the "c"
continous mode command which toggles the output on or
off.

The duplicate comment documenting the command usage
has been removed as it was already starting to drift from the
usage command output.

Usage information is shuffled so that the new commands are
before the "u" usage command.

Tests:
 - Run updated unit tests
 - Enasure that on gauge start up no additional output is
   present
 - Enable continuous mode and verify output matches analog
   inputs to the tank
 - Turn off continous mode and verify output is quiet
 - Check that when in continuous mode is enabled that going
   to program mode stops output
 - Verify usage output