Indicate an error with the tank input being out of range
This checks the tank input ADC to see if it is going full scale.
This indicates that there is an error with the sender
resistance sensing. The error is indicated to the user by
flashing the low fuel indicator.

Slightly odd side-effect is that in continuous mode the sensor
logging output stops while there is an error. Outputting text
at the same time as flashing the LEDs is difficult due to dumb
loop used.

Tests:
 - Run updated unit tests which check the return result of
   RunGauge()
 - Run the gauge on the target HW and force ADC errors
   using a 0-5V pot  (expected input 0-1.024V)
 - Run in normal and ocntinuous logging modes and verify
   error flashes are displayed