Enable the system watchdog
This enables the system watchdog on the PIC MCU. The default 2 second timeout is used as this should be long
enough to cover and eventuality in gauge operation.

The watchdog is strobed each time around the main loop
which should be at most every 1ms. The MCC code generator
inserts other watchdog strobes in the flash programming
(unused) and UART reading.

Should the watchdog trigger a message is printed to the
console. It doesn't seem serious enough to warrant strobing
the low fuel LED.

According to the XC8 compiler manual the -mresetbits
option is required to detect the watchdog timeout on start
up. Testing reveals no difference in operation. The compiler
option is turned on anyway.

Tests:
 - Run the gauge through all operating modes including
   generating large amounts of serial output which could take
   extended periods of time.
 - Hack the line too long error to force a 3 second timeout.
   Check that the watchdog triggers and that this is logged in
   the serial output