Fix a sensor error not being reported on startup
This fixes an issue found where the board starts up without a
sensor connected but the low fuel LED would not flash to
indicate the error condition. Upon inspection it appears that
the errorCount is not being initialised. This causes the
errorCode value to contain random values which can take a
long time in the error condition before it starts flashing.

Tests:
 - Power the unit on with no sender unit conected and verify
   that the low fuel LED flashses at 1Hz as expected