Add a function to map from an input map to an output map
This change adds a function (with tests) which can take two
maps to relate an input value to an output value in a
different number space. This will eventually allow a non-
linear input number to be mapped to a linear tank level
value with a full-scale of 0x0000 to 0xffff. It can also be used
to map a linear tank level to a non-linear gauge output value.

No attempt is made to integrate with the rest of the gauge.

This function is a general version of the MapValue() function
which assumes a linear input map with steps of 0x2000 per bin.

Tests:
 - Run updated unit tests