Replace the over-optimised implementation of MapValue()
This replaces the over-optimised implementation of
MapValue() which makes a static assumption that the top 3
bits can be used to identify the output map bin by one which
uses an explicit input map. This is much more accurate.

Update all the tests across the system to remove the fudging
required because the top value cannot fit 0x10000 in a
uint16_t.

Tests:
 - Run updated unit tests having checked each value change
   manually.