Change input mapping direction to be more logical
This change uses the new mapping function to allow the tank
input values to be mapped to the linear actual values
accurately and to use the full scale.

The change removes the original MapValue function and
replaces it with the new input/output map function. This
requires extensive refactoring of the original MapperTest
unit and renaming in InputMapperTest. The logic and values
in each test is unchanged.

The change in mapping semantics results in re-applying a
fudge to the gauge mapping tests due to the loss of precision
in the upper bin e000-ffff.

Tests:
 - Run updated unit tests