Fix numerical precision issues when running on a 8-bit PIC
This fixes a problem when running on an 8-bit PIC where the calcuations used in interplotation are not carried out using a 32-bit signed integer as required but as a 16-bit unsigned value. Explicitly casting the first use in each expression ensures that the correct type is used for the intermediate value.

Tests:
 - Run updated unit test with a real-world test value
 - Build and run on a PIC and verify in the debugger that the
   calculation is operating the same as the unit test for the
   same input value