Properly escape the git version string
This fixes an error in the stringify macro used to include the
git version that resulted in the string GIT_VERSION being
displayed rather than the actual hash.

The quoting in the define passed to the compiler has been
removed as this caused differences between the Linux and
Windows builds.